[TOOLS-2383] Structural Review Changes
* Organize classes within packages;
