+++ kpi-parent/kpi-cycle-time/src/main/java/objective/taskboard/followup/kpi/cycletime/CycleTimeController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import objective.taskboard.auth.authorizer.permission.ProjectDashboardOperationalPermission;
 import objective.taskboard.followup.kpi.KpiLevel;
-import objective.taskboard.followup.kpi.touchTime.KpiValidationException;
+import objective.taskboard.followup.kpi.KpiValidationException;
 import objective.taskboard.jira.ProjectService;
 
 @RestController
