+++ kpi-parent/kpi-touch-time/src/main/java/objective/taskboard/followup/kpi/touchtime/TouchTimeKpiController.java
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import objective.taskboard.auth.authorizer.permission.ProjectDashboardOperationalPermission;
 import objective.taskboard.followup.kpi.KpiLevel;
-import objective.taskboard.followup.kpi.touchTime.KpiValidationException;
+import objective.taskboard.followup.kpi.KpiValidationException;
 import objective.taskboard.jira.ProjectService;
 
 @RestController
